+++ src/lib/viz/dataview/histogram/HistogramDataViewImpl.ts
@@ -1,6 +1,6 @@
 import { isVariableDefined } from '@/core/utils/variables';
 import { isDefined } from '@/viz/utils/object';
-import { AggregationType, aggregate } from '@/data/operations/aggregation/aggregation';
+import { AggregationType, aggregateValues } from '@/data/operations/aggregation';
 import { DataViewMode, DataViewCalculation } from '../mode/DataViewMode';
 import { DataViewImpl } from '../DataViewImpl';
 import { CartoDataViewError, dataViewErrorTypes } from '../DataViewError';
@@ -31,21 +31,8 @@ export class HistogramDataViewImpl extends DataViewImpl<HistogramDataViewData> {
     const columnName = this.column;
 
     try {
-      const features = (await dataviewLocal.getSourceData({
-        ...options,
-        aggregationOptions: {
-          numeric: [
-            {
-              column: this.column,
-              operations: [this.operation]
-            }
-          ]
-        }
-      })) as Record<string, number>[];
-
-      const sortedFeatures = features
-        .map(feature => getFeatureValue(feature, aggregatedColumnName, columnName).featureValue)
-        .sort((a, b) => a - b);
+      const features = (await dataviewLocal.getSourceData(options)) as Record<string, number>[];
+      const sortedFeatures = features.map(feature => feature[this.column]).sort((a, b) => a - b);
 
       const startValue = start ?? Math.min(...sortedFeatures);
       const endValue = end ?? Math.max(...sortedFeatures);
@@ -92,9 +79,9 @@ export class HistogramDataViewImpl extends DataViewImpl<HistogramDataViewData> {
           start: binContainer.start,
           end: binContainer.end,
           value: binContainer.value,
-          min: aggregate(binContainer.values, AggregationType.MIN),
-          max: aggregate(binContainer.values, AggregationType.MAX),
-          avg: aggregate(binContainer.values, AggregationType.AVG),
+          min: aggregateValues(binContainer.values, AggregationType.MIN).result,
+          max: aggregateValues(binContainer.values, AggregationType.MAX).result,
+          avg: aggregateValues(binContainer.values, AggregationType.AVG).result,
           normalized: binContainer.values.length / features.length
         };
       });