+++ src/lib/viz/dataview/mode/DataViewLocal.ts
@@ -1,39 +1,49 @@
-import { Layer, GeoJSONSource } from '@/viz';
-import { AggregationType, aggregate } from '@/data/operations/aggregation/aggregation';
+import { Layer } from '@/viz';
+import { DATA_CHANGED_EVENT } from '@/viz/layer/Layer';
+import { AggregationType, aggregateValues } from '@/data/operations/aggregation';
 import { groupValuesByColumn } from '@/data/operations/grouping';
 import { castToNumberOrUndefined } from '@/core/utils/number';
 import { ColumnFilters, SpatialFilters } from '@/viz/filters/types';
 import { DataViewMode } from './DataViewMode';
 
 export class DataViewLocal extends DataViewMode {
-  private useViewport = true;
+  private useViewport = false;
 
-  constructor(dataSource: Layer, column: string, useViewport = true) {
+  constructor(dataSource: Layer, column: string) {
     super(dataSource, column);
-
-    this.useViewport = useViewport;
     this.bindEvents();
   }
 
-  public async getSourceData(columns: string[] = [], options: { excludedFilters?: string[] } = {}) {
-    if (!columns.includes(this.column)) {
-      columns.push(this.column);
-    }
-
-    const { excludedFilters = [] } = options;
+  public async getSourceData(
+    options: {
+      excludedFilters?: string[];
+      aggregationOptions?: {
+        dimension?: string[];
+        numeric?: { column: string; operations: string[] }[];
+      };
+    } = {}
+  ) {
+    const { excludedFilters = [], aggregationOptions } = options;
 
     if (this.useViewport) {
-      await (this.dataSource as Layer).addSourceField(this.column);
+      await (this.dataSource as Layer).addAggregationOptions(
+        aggregationOptions?.numeric,
+        aggregationOptions?.dimension
+      );
+
       return (this.dataSource as Layer).getViewportFeatures(excludedFilters);
     }
 
     // is GeoJSON Layer
     if (this.dataSource instanceof Layer) {
-      return (this.dataSource.source as GeoJSONSource).getFeatures(excludedFilters);
+      await this.dataSource.addSourceField(this.column);
+    } else {
+      this.dataSource.addField(this.column);
     }
 
-    // is GeoJSON Source
-    return (this.dataSource as GeoJSONSource).getFeatures(excludedFilters);
+    return this.useViewport
+      ? (this.dataSource as Layer).getViewportFeatures(excludedFilters)
+      : this.dataSource.getFeatures(excludedFilters);
   }
 
   public async groupBy(
@@ -41,7 +51,7 @@ export class DataViewLocal extends DataViewMode {
     operation: AggregationType,
     options: { excludedFilters: string[] }
   ) {
-    const sourceData = await this.getSourceData([operationColumn || this.column], options);
+    const sourceData = await this.getSourceData(options);
     const { groups, nullCount } = groupValuesByColumn(
       sourceData,
       operationColumn || this.column,
@@ -60,15 +70,14 @@ export class DataViewLocal extends DataViewMode {
     }
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars, class-methods-use-this
-  public setSpatialFilter(_spatialFilter: SpatialFilters) {
-    return undefined;
+  public setSpatialFilter(spatialFilter: SpatialFilters) {
+    this.useViewport = spatialFilter === 'viewport';
   }
 
   private bindEvents() {
     this.registerAvailableEvents(['dataUpdate', 'error']);
 
-    this.dataSource.on('viewportLoad', () => {
+    this.dataSource.on(DATA_CHANGED_EVENT, () => {
       this.onDataUpdate();
     });
 
@@ -94,7 +103,7 @@ function createCategory(name: string, data: number[], operation: AggregationType
 
   return {
     name,
-    value: aggregate(categoryValues, operation)
+    value: aggregateValues(categoryValues, operation).result
   };
 }