+++ src/lib/viz/source/GeoJSONSource.ts
@@ -1,6 +1,6 @@
 import { GeoJSON, Feature, GeoJsonGeometryTypes } from 'geojson';
 import { uuidv4 } from '@/core/utils/uuid';
-import { aggregate, AggregationType } from '@/data/operations/aggregation/aggregation';
+import { aggregateValues, AggregationType } from '@/data/operations/aggregation';
 import { FiltersCollection } from '../filters/FiltersCollection';
 import { FunctionFilterApplicator } from '../filters/FunctionFilterApplicator';
 import { ColumnFilters } from '../filters/types';
@@ -18,7 +18,7 @@ import { sourceErrorTypes, SourceError } from '../errors/source-error';
 import { selectPropertiesFrom } from '../utils/object';
 
 interface GeoJSONSourceProps extends SourceProps {
-  data: GeoJSON;
+  data: Feature[];
 }
 
 export const DEFAULT_GEOM = 'Point';
@@ -75,7 +75,7 @@ export class GeoJSONSource extends Source {
       return true;
     }
 
-    this._props = { type: 'GeoJSONLayer', data: this._geojson };
+    this._props = { type: 'GeoJSONLayer', data: getFeatures(this._geojson) };
     this._metadata = this._buildMetadata();
 
     this.needsInitialization = false;
@@ -90,6 +90,10 @@ export class GeoJSONSource extends Source {
   removeFilter(filterId: string) {
     this.filtersCollection.removeFilter(filterId);
     this.emit('filterChange');
+  }
+
+  isEmpty() {
+    return getFeatures(this._geojson).length === 0;
   }
 
   private _buildMetadata() {
@@ -184,10 +188,10 @@ export class GeoJSONSource extends Source {
 
     // eslint-disable-next-line no-restricted-syntax
     for (const [propName, values] of Object.entries(this._numericFieldValues)) {
-      const min = aggregate(values, AggregationType.MIN);
-      const max = aggregate(values, AggregationType.MAX);
-      const avg = aggregate(values, AggregationType.AVG);
-      const sum = aggregate(values, AggregationType.SUM);
+      const min = aggregateValues(values, AggregationType.MIN).result;
+      const max = aggregateValues(values, AggregationType.MAX).result;
+      const avg = aggregateValues(values, AggregationType.AVG).result;
+      const sum = aggregateValues(values, AggregationType.SUM).result;
       const sample = createSample(values);
 
       numericStats.push({