+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
 - Add `beforeLayerId` and `afterLayerId` options to `carto.viz.Layer.addTo` method, to customize layer position
 - Widget synchronization when using remote dataviews ([#76](https://github.com/CartoDB/web-sdk/pull/76/))
+- `icon` style helper 
+- New data events in Layer: `dataReady` and `dataChanged` ([#84](https://github.com/CartoDB/web-sdk/pull/84))
 
 ### Changed
 - Remove d3.format option from `carto.viz.Popup`
@@ -17,5 +19,3 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [1.0.0-alpha.0] 2020-07-06
 - First release
-
-